0.0.7 (#10)
* Add dplyr rows verbs

* Add tidyr.chop/unchop

* Add tidyr.pack/unpack

* Allow mixed numbering for tibble construction

* Add tidyr.nest/unnest

* Add tidyr.expand/expand_grid

* Add tidyr.complete

* Move tidyr.drop_na

* Move tidyr.extract

* Move tidyr.fill

* Move and fix tidyr.pivot_longer

* Move and fix tidyr.pivot_wider

* Move and fix tidyr.separate/separate_rows/uncount

* Move and fix tidyr.unite

* Move and fix tidyr.replace_na

* Update notebook for tidyr.full_seq

* 0.0.7

* Update CHANGELOG

+++ datar/base/constants.py
@@ -18,6 +18,6 @@ letters = numpy.array(list(ascii_letters[:26])) # pylint: disable=invalid-name
 LETTERS = numpy.array(list(ascii_letters[26:]))
 
 NA_character_ = f"<NA_{uuid.uuid4()}_>"
-NA_integer_ = numpy.random.randint(numpy.iinfo(numpy.int64).max)
+NA_integer_ = numpy.random.randint(numpy.iinfo(numpy.int32).max)
 NA_real_ = NA
 NA_compex_ = complex(NA, NA)
